requirements.txt
anyio==3.7.1
APScheduler==3.10.3
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
cryptography==41.0.3
distlib==0.3.7
Django==4.2.3
django-cleanup==8.0.0
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.12.4
future==0.18.3
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
numpy==1.25.2
pbr==5.11.1
Pillow==10.0.0
platformdirs==3.10.0
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==20.4
pytz==2023.3
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
stevedore==5.1.0
tzdata==2023.3
tzlocal==5.0.1
virtualenv==20.24.5
virtualenv-clone==0.5.7
